Move prepare_url func from core.emails module to core.utils
