+++ saleor/core/emails.py
@@ -1,5 +1,3 @@
-from urllib.parse import urlsplit
-
 from django.contrib.sites.models import Site
 
 
@@ -11,10 +9,3 @@ def get_email_context():
         "site_name": site.name,
     }
     return send_email_kwargs, email_template_context
-
-
-def prepare_url(params: str, redirect_url: str) -> str:
-    """Add params to redirect url."""
-    split_url = urlsplit(redirect_url)
-    split_url = split_url._replace(query=params)
-    return split_url.geturl()
